Install the latest version of the template instead of the dev version
instead of calling git, use "composer create-project" to clone the latest
version of the template, but without executing any code after that.

This gives us more predictable results, as random dev versions of the template
wont be used, only the latest tagged on.

Also we find the package via packagist, like everything else and don't need to
hardcode the github repo.